tapo-proxy/requirements.txt
python-kasa>=0.9.0
flask>=3.0.0
